refactor Bin class tests
