+++ src/test/java/org/cicirello/ibp/BinTests.java
@@ -219,146 +219,6 @@ public class BinTests {
   }
 
   @Test
-  public void testBinSortIncreasing() {
-    Item[][] alreadySorted = {
-      {new Item("A", 5)},
-      {new Item("A", 5), new Item("B", 10)},
-      {new Item("A", 5), new Item("B", 10), new Item("C", 15)},
-      {new Item("A", 5), new Item("B", 10), new Item("C", 15), new Item("D", 20)},
-      {new Item("A", 5), new Item("B", 10), new Item("C", 15), new Item("D", 20), new Item("E", 25)}
-    };
-    Item[][] reversed = {
-      {new Item("A", 5)},
-      {new Item("B", 10), new Item("A", 5)},
-      {new Item("C", 15), new Item("B", 10), new Item("A", 5)},
-      {new Item("D", 20), new Item("C", 15), new Item("B", 10), new Item("A", 5)},
-      {new Item("E", 25), new Item("D", 20), new Item("C", 15), new Item("B", 10), new Item("A", 5)}
-    };
-    Item[][] jumbled = {
-      {new Item("C", 15), new Item("A", 5), new Item("D", 20), new Item("B", 10)},
-      {new Item("C", 15), new Item("A", 5), new Item("E", 25), new Item("D", 20), new Item("B", 10)}
-    };
-
-    for (int i = 0; i < alreadySorted.length; i++) {
-      Bin b = new Bin("Bin Name", -1);
-      for (int j = 0; j < alreadySorted[i].length; j++) {
-        assertTrue(b.add(alreadySorted[i][j]));
-      }
-      ArrayList<Item> contents = b.getContents();
-      for (int j = 0; j < alreadySorted[i].length; j++) {
-        assertEquals(alreadySorted[i][j], contents.get(j));
-      }
-      b.sort(false);
-      contents = b.getContents();
-      for (int j = 0; j < alreadySorted[i].length; j++) {
-        assertEquals(alreadySorted[i][j], contents.get(j));
-      }
-    }
-
-    for (int i = 0; i < reversed.length; i++) {
-      Bin b = new Bin("Bin Name", -1);
-      for (int j = 0; j < reversed[i].length; j++) {
-        assertTrue(b.add(reversed[i][j]));
-      }
-      ArrayList<Item> contents = b.getContents();
-      for (int j = 0; j < reversed[i].length; j++) {
-        assertEquals(reversed[i][j], contents.get(j));
-      }
-      b.sort(false);
-      contents = b.getContents();
-      for (int j = 0; j < alreadySorted[i].length; j++) {
-        assertEquals(alreadySorted[i][j], contents.get(j));
-      }
-    }
-
-    for (int i = 0; i < jumbled.length; i++) {
-      Bin b = new Bin("Bin Name", -1);
-      for (int j = 0; j < jumbled[i].length; j++) {
-        assertTrue(b.add(jumbled[i][j]));
-      }
-      ArrayList<Item> contents = b.getContents();
-      for (int j = 0; j < jumbled[i].length; j++) {
-        assertEquals(jumbled[i][j], contents.get(j));
-      }
-      b.sort(false);
-      contents = b.getContents();
-      for (int j = 0; j < alreadySorted[i + 3].length; j++) {
-        assertEquals(alreadySorted[i + 3][j], contents.get(j));
-      }
-    }
-  }
-
-  @Test
-  public void testBinSortDecreasing() {
-    Item[][] reversed = {
-      {new Item("A", 5)},
-      {new Item("A", 5), new Item("B", 10)},
-      {new Item("A", 5), new Item("B", 10), new Item("C", 15)},
-      {new Item("A", 5), new Item("B", 10), new Item("C", 15), new Item("D", 20)},
-      {new Item("A", 5), new Item("B", 10), new Item("C", 15), new Item("D", 20), new Item("E", 25)}
-    };
-    Item[][] alreadySorted = {
-      {new Item("A", 5)},
-      {new Item("B", 10), new Item("A", 5)},
-      {new Item("C", 15), new Item("B", 10), new Item("A", 5)},
-      {new Item("D", 20), new Item("C", 15), new Item("B", 10), new Item("A", 5)},
-      {new Item("E", 25), new Item("D", 20), new Item("C", 15), new Item("B", 10), new Item("A", 5)}
-    };
-    Item[][] jumbled = {
-      {new Item("C", 15), new Item("A", 5), new Item("D", 20), new Item("B", 10)},
-      {new Item("C", 15), new Item("A", 5), new Item("E", 25), new Item("D", 20), new Item("B", 10)}
-    };
-
-    for (int i = 0; i < alreadySorted.length; i++) {
-      Bin b = new Bin("Bin Name", -1);
-      for (int j = 0; j < alreadySorted[i].length; j++) {
-        assertTrue(b.add(alreadySorted[i][j]));
-      }
-      ArrayList<Item> contents = b.getContents();
-      for (int j = 0; j < alreadySorted[i].length; j++) {
-        assertEquals(alreadySorted[i][j], contents.get(j));
-      }
-      b.sort(true);
-      contents = b.getContents();
-      for (int j = 0; j < alreadySorted[i].length; j++) {
-        assertEquals(alreadySorted[i][j], contents.get(j));
-      }
-    }
-
-    for (int i = 0; i < reversed.length; i++) {
-      Bin b = new Bin("Bin Name", -1);
-      for (int j = 0; j < reversed[i].length; j++) {
-        assertTrue(b.add(reversed[i][j]));
-      }
-      ArrayList<Item> contents = b.getContents();
-      for (int j = 0; j < reversed[i].length; j++) {
-        assertEquals(reversed[i][j], contents.get(j));
-      }
-      b.sort(true);
-      contents = b.getContents();
-      for (int j = 0; j < alreadySorted[i].length; j++) {
-        assertEquals(alreadySorted[i][j], contents.get(j));
-      }
-    }
-
-    for (int i = 0; i < jumbled.length; i++) {
-      Bin b = new Bin("Bin Name", -1);
-      for (int j = 0; j < jumbled[i].length; j++) {
-        assertTrue(b.add(jumbled[i][j]));
-      }
-      ArrayList<Item> contents = b.getContents();
-      for (int j = 0; j < jumbled[i].length; j++) {
-        assertEquals(jumbled[i][j], contents.get(j));
-      }
-      b.sort(true);
-      contents = b.getContents();
-      for (int j = 0; j < alreadySorted[i + 3].length; j++) {
-        assertEquals(alreadySorted[i + 3][j], contents.get(j));
-      }
-    }
-  }
-
-  @Test
   public void testBinRemoveAll() {
     Item[][] items = {
       {new Item("A", 5)},
